api: refactored api to use delegate
