+++ src/backend/jag-efiling-api/pom.xml
@@ -15,9 +15,9 @@
 	<description>Demo project for Spring Boot</description>
 
 	<properties>
-		<default-package>ca.bc.gov.open</default-package>
 		<java.version>1.8</java.version>
 		<org.mapstruct.version>1.3.1.Final</org.mapstruct.version>
+		<default-package>ca.bc.gov.open.jagefilingapi</default-package>
 	</properties>
 	<dependencies>
 
@@ -205,6 +205,7 @@
 							<generatorName>spring</generatorName>
 							<configOptions>
 								<dateLibrary>joda</dateLibrary>
+								<delegatePattern>true</delegatePattern>
 							</configOptions>
 							<library>spring-boot</library>
 							<apiPackage>${default-package}.api</apiPackage>
