+++ src/backend/jag-efiling-api/src/test/java/ca/bc/gov/open/jagefilingapi/document/SubmissionApiImplTest.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.open.jagefilingapi.document;
 
-import ca.bc.gov.open.api.model.*;
+
+import ca.bc.gov.open.jagefilingapi.api.model.*;
 import ca.bc.gov.open.jagefilingapi.config.NavigationProperties;
 import ca.bc.gov.open.jagefilingapi.fee.FeeService;
 import ca.bc.gov.open.jagefilingapi.fee.models.Fee;
@@ -26,7 +27,9 @@ import java.util.Collections;
 import java.util.Optional;
 import java.util.UUID;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
